Implement Google OAuth

+++ src/main/java/org/example/pactimemultiplayer/service/LobbyService.java
@@ -7,6 +7,7 @@ import org.example.pactimemultiplayer.entity.Lobby;
 import org.example.pactimemultiplayer.entity.Player;
 import org.example.pactimemultiplayer.entity.GameMode;
 import org.example.pactimemultiplayer.events.LobbyUpdatedEvent;
+import org.example.pactimemultiplayer.exception.*;
 import org.example.pactimemultiplayer.repository.LobbyRepository;
 import org.example.pactimemultiplayer.repository.GameModeRepository;
 import lombok.RequiredArgsConstructor;
@@ -21,7 +22,6 @@ import java.util.List;
 public class LobbyService {
     private final LobbyRepository lobbyRepository;
     private final GameModeRepository gameModeRepository;
-    private final PlayerService playerService;
     private final GameFlowService gameFlowService;
     private final ApplicationEventPublisher eventPublisher;
 
@@ -32,38 +32,32 @@ public class LobbyService {
                 .toList();
     }
 
-    public LobbyDto createLobby(CreateLobbyDto dto) {
+    public LobbyDto createLobby(CreateLobbyDto dto, Player player) {
         String code = generateUniqueCode();
-        Player host = playerService.findOrCreate(dto.getHostUsername());
 
         Lobby lobby = new Lobby();
         lobby.setCode(code);
-        lobby.setHost(host);
+        lobby.setHost(player);
         lobby.setName(dto.getName());
         lobby.setStarted(false);
 
         GameMode mode = gameModeRepository.findByShortName(dto.getGmShortName())
-                .orElseThrow(() -> new RuntimeException("Game mode not found"));
+                .orElseThrow(InvalidGameModeException::new);
 
         lobby.setMode(mode);
-        lobby.getPlayers().add(host);
+        lobby.getPlayers().add(player);
         lobbyRepository.save(lobby);
 
         return LobbyDto.from(lobby);
     }
 
-    public LobbyDto joinLobby(JoinLobbyDto dto) {
-        Lobby lobby = lobbyRepository.findById(dto.getLobbyCode())
-                .orElseThrow(() -> new RuntimeException("Lobby not found"));
+    public LobbyDto joinLobby(JoinLobbyDto dto, Player player) {
+        Lobby lobby = lobbyRepository.findById(dto.getLobbyCode()).orElseThrow(LobbyNotFoundException::new);
 
-        if (lobby.isStarted()) {
-            throw new RuntimeException("Lobby already started");
-        }
-        if (lobby.getPlayers().size() >= Constants.LOBBY_MAX_PLAYERS) {
-            throw new RuntimeException("Lobby is full");
-        }
+        if (lobby.getPlayers().contains(player)) throw new PlayerIsAlreadyInLobbyException();
+        if (lobby.isStarted()) throw new LobbyStartedException();
+        if (lobby.getPlayers().size() >= Constants.LOBBY_MAX_PLAYERS) throw new LobbyFullException();
 
-        Player player = playerService.findOrCreate(dto.getUsername());
         lobby.getPlayers().add(player);
         lobbyRepository.save(lobby);
 
@@ -71,21 +65,22 @@ public class LobbyService {
     }
 
     @Transactional
-    public LobbyDto leaveLobby(LeaveLobbyDto dto) {
-        Lobby lobby = removePlayer(dto.getLobbyCode(), dto.getUsername());
+    public LobbyDto leaveLobby(LeaveLobbyDto dto, Player player) {
+        Lobby lobby = removePlayer(dto.getLobbyCode(), player.getId());
         return LobbyDto.from(lobby);
     }
 
-    public boolean deleteLobby(DeleteLobbyDto dto) {
-        Lobby lobby = lobbyRepository.findById(dto.getLobbyCode())
-                .orElseThrow(() -> new RuntimeException("Lobby not found"));
-        lobbyRepository.delete(lobby);
+    public void deleteLobby(DeleteLobbyDto dto, Player player) {
+        Lobby lobby = lobbyRepository.findById(dto.getLobbyCode()).orElseThrow(LobbyNotFoundException::new);
+        if (!player.getId().equals(lobby.getHost().getId())) throw new PlayerIsNotTheHostException();
 
-        return true;
+        lobbyRepository.delete(lobby);
     }
 
-    public LobbyDto startLobby(StartLobbyDto dto) {
-        Lobby lobby = lobbyRepository.findById(dto.getCode()).orElseThrow();
+    public LobbyDto startLobby(StartLobbyDto dto, Player player) {
+        Lobby lobby = lobbyRepository.findById(dto.getLobbyCode()).orElseThrow(LobbyNotFoundException::new);
+        if (!player.getId().equals(lobby.getHost().getId())) throw new PlayerIsNotTheHostException();
+
         lobby.setStarted(true);
         lobbyRepository.save(lobby);
 
@@ -94,17 +89,13 @@ public class LobbyService {
     }
 
     @Transactional
-    public Lobby removePlayer(String code, String username) {
-        Lobby lobby = lobbyRepository.findWithPlayers(code)
-                .orElseThrow(() -> new RuntimeException("Lobby not found"));
+    public Lobby removePlayer(String code, String id) {
+        Lobby lobby = lobbyRepository.findWithPlayers(code).orElseThrow(LobbyNotFoundException::new);
 
-        boolean removed = lobby.getPlayers()
-                .removeIf(p -> p.getUsername().equals(username));
+        boolean removed = lobby.getPlayers().removeIf(p -> p.getId().equals(id));
+        if (!removed) throw new PlayerNotInLobbyException();
 
-        if (removed) {
-            eventPublisher.publishEvent(new LobbyUpdatedEvent(code));
-        }
-
+        eventPublisher.publishEvent(new LobbyUpdatedEvent(code));
         return lobby;
     }
